collapse "prettier" task into "npm-scripts"
