+++ lib/tasks/npm-scripts.js
@@ -1,6 +1,7 @@
 /* @flow */
 'use strict';
 
+const { intersects } = require('semver');
 const omit = require('lodash.omit');
 const path = require('path');
 const updateJsonFile = require('update-json-file');
@@ -26,10 +27,25 @@ function ava(pkg) {
 }
 
 const fmt = () => ({
-  fmt: 'npm run sort-package-json',
+  fmt: 'npm run sort-package-json && npm run prettier',
+  prettier:
+    'npx -q prettier --loglevel=warn --write "**/*.{css,js,json,jsx,md,less,scss,ts,tsx}"',
   pretest: 'npm run fmt',
   'sort-package-json': 'npx -q sort-package-json',
 });
+
+const prettier = pkg => ({
+  singleQuote: true,
+  trailingComma: trailingComma(pkg),
+});
+
+function trailingComma(pkg /* : Object */) /* : 'all' | 'es5' */ {
+  // Node.js 8.x supports "all", otherwise use "es5"
+  if (!pkg.engines.node || intersects(pkg.engines.node, '<8')) {
+    return 'es5';
+  }
+  return 'all';
+}
 
 function fn({ cwd } /* : TaskOptions */) {
   return updateJsonFile(
@@ -38,8 +54,12 @@ function fn({ cwd } /* : TaskOptions */) {
       // a temporary copy of pkg that requires fewer checks
       const safePkg = {
         devDependencies: pkg.devDevendencies || {},
+        engines: pkg.engines || {},
         ...pkg,
       };
+
+      // TODO: ought to consider a better place for this?
+      pkg.prettier = prettier(safePkg);
 
       pkg.scripts = {
         ...ava(safePkg),
