+++ lib/tasks/npm-uninstall.js
@@ -12,8 +12,12 @@ const LABEL = 'npm packages uninstalled';
 /* :: import type { TaskOptions } from '../types.js' */
 
 async function fn({ cwd } /* : TaskOptions */) {
-  const pkgs = ['fixpack'];
-  await execa('npm', ['uninstall', ...pkgs], { cwd });
+  const pkgs = [
+    'fixpack', // use sort-package-json now instead
+    'prettier', // run this via npx instead of as a devDep
+  ];
+  await execa('npm', ['uninstall', '--save', ...pkgs], { cwd });
+  await execa('npm', ['uninstall', '--save-dev', ...pkgs], { cwd });
 }
 
 module.exports = {
